% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maths.R
\name{baseline_ftir}
\alias{baseline_ftir}
\title{Baseline FTIR}
\usage{
baseline_ftir(ftir, method = "modpolyfit", ...)
}
\arguments{
\item{ftir}{A data.frame in long format with columns `sample_id`,
  `wavenumber`, and `absorbance`. The `absorbance` column may be replaced by
  a `transmittance` column for transmittance plots.

  Un data.frame au format long avec les colonnes `sample_id`, `wavenumber`,
  et `absorbance`. La colonne `absorbance` peut être remplacée par une
  colonne `transmittance` pour les tracés de transmission.}

\item{method}{A method from [baseline::baseline()]. For FTIR data, best
  results are achieved by selecting either `modpolyfit`, `peakDetection` or
  `rfbaseline`.

  Une méthode de [baseline::baseline()]. Pour les données IRTF, les meilleurs
  résultats sont obtenus en sélectionnant soit `modpolyfit`, soit
  `peakDetection`, soit `rfbaseline`.}

\item{...}{Additional parameters required by specific methods in
  [baseline::baseline()].

  Paramètres supplémentaires requis par certaines méthodes de
  [baseline::baseline()].}
}
\value{
A FTIR spectral data.frame with baseline corrected intensity column.

Un data.frame contenant le spectre IRTF corrigé de la ligne de base
}
\description{
Correct the baseline of an FTIR spectrum using one of the
techniques available in the [baseline::baseline()] package.

Corrigez la ligne de base d'un spectre IRTF en utilisant l'une des techniques
disponibles dans le package [baseline::baseline()].
}
\examples{
# Load the isopropanol sample spectrum
ftir_data <- sample_spectra[
  sample_spectra$sample_id == "isopropanol",
]

# Apply baseline correction using the default 'modpolyfit' method
ftir_baselined_modpoly <- baseline_ftir(ftir_data)

# Apply baseline correction using the 'lowpass' method
ftir_baselined_lowpass <- baseline_ftir(ftir_data, method = "lowpass")

# --- Optional: Visualize the results ---
\dontrun{
  plot_ftir(ftir_baselined_modpoly, plot_title = "ModPoly Baselined FTIR")

  plot_ftir(ftir_baselined_lowpass, plot_title = "Lowpass Baselined FTIR")
}
}
\references{
* Kristian Hovde Liland, Trygve Almøy, Bjørn-Helge Mevik (2010) Optimal Choice of Baseline Correction for Multivariate Calibration of Spectra, Applied Spectroscopy 64, pp. 1007-1016. doi:10.1366/000370210792434350
* Chad A. Lieber and Anita Mahadevan-Jansen (2003) Automated Method for Subtraction of Fluorescence from Biological Raman Spectra, Applied Spectroscopy 57, pp. 1363-1367. doi:10.1366/000370203322554518
* Kevin R. Coombes et al. (2003) Quality control and peak finding for proteomics data collected from nipple aspirate fluid by surface-enhanced laser desorption and ionization. Clinical Chemistry 49, pp. 1615-1623. doi:10.1373/49.10.1615
* Andreas F. Ruckstuhl, Matthew P. Jacobson, Robert W. Field, James A. Dodd (2001) Baseline subtraction using robust local regression estimation. Journal of Quantitative Spectroscopy and Radiative Transfer 68, pp.. 179-193. doi:10.1016/S0022-4073(00)00021-2
* Xianchun Shen et al. (2018) Applied Optics 57 pp. 5794-5799 doi:10.1364/AO.57.0057947
}
\seealso{
[baseline::baseline()]
}
