% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{c.ftir}
\alias{c.ftir}
\title{Vector FTIR}
\usage{
\method{c}{ftir}(...)
}
\arguments{
\item{...}{multiple objects to put in a single object}
}
\value{
a single ftir object
}
\description{
Produces an unelementable ftir spectra object
}
