% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compress_region.R
\name{compress_trans}
\alias{compress_trans}
\title{Compression Transformation}
\usage{
compress_trans(intercept, reducer)
}
\arguments{
\item{intercept}{the wavenumber at which compression starts}

\item{reducer}{the rate of compression (larger numbers are more compressed)}
}
\description{
used in compress_low_energy()
}
\references{
From https://stackoverflow.com/a/43321639 & https://stackoverflow.com/a/78360580
}
\keyword{internal}
