% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{-.ftir}
\alias{-.ftir}
\title{Subtraction of FTIR objects}
\usage{
\method{-}{ftir}(x, y)
}
\arguments{
\item{x}{First item to subtract. Either `x` or `y` must be an object of `class() == "ftir"`, and `y` or `x` must be scalar}

\item{y}{Second item to subtract. Either `x` or `y` must be an object of `class() == "ftir"`, and `y` or `x` must be scalar}
}
\value{
A data.frame containing the adjusted FTIR spectra.

  Un data.frame contenant les spectres IRTF ajustee.
}
\description{
Subtract a numeric (scalar) value from a FTIR spectra. Note: always operates as FTIR - x regardless of the order of values provided.
}
\seealso{
[subtract_scalar_value()]
}
