% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulations.R
\name{-.gg}
\alias{-.gg}
\title{Add GGPlot Layer below others}
\usage{
\method{-}{gg}(plot, layer)
}
\arguments{
\item{plot}{The plot to which the layer should be added.

  Le tracé auquel le calque doit être ajouté.}

\item{layer}{The layer to add to the plot.

La couche à ajouter au tracé.}
}
\value{
The ggplot2 `plot` object, with layer `layer` added underneath.

  L'objet ggplot2 `plot`, avec le calque `layer` ajouté en dessous.
}
\description{
Inserts a layer *underneath* the existing layers on a ggplot
  object simply by calling `plot` - `layer` instead of the usual `plot` +
  `layer`.

  Insère un calque *sous* les calques existants sur un objet ggplot
  simplement en appelant `plot` - `layer` au lieu de l'habituel `plot` +
  `layer`.
}
\references{
From https://stackoverflow.com/a/64011534
}
\keyword{internal}
