% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak-fit.R
\name{fit_peak_df}
\alias{fit_peak_df}
\title{Fitted Peaks Data.Frame}
\usage{
fit_peak_df(fitted_peaks)
}
\arguments{
\item{fitted_peaks}{An object from [fit_peaks()].

  Un objet de [fit_peaks()].}
}
\value{
A data.frame of peak properties.

  Un data.frame des propriétés des pics.
}
\description{
Reformat the [fit_peaks()] object to a data.frame of peak
specifications.

Reformater l'objet [fit_peaks()] en un data.frame de spécifications de pics.
}
\examples{
# Load the isopropanol sample spectrum from the PlotFTIR package
ftir_data <- PlotFTIR::sample_spectra[
  PlotFTIR::sample_spectra$sample_id == "isopropanol",
]

# Choose a subset of the data (reducing run time)
ftir_data <- ftir_data[
  ftir_data$wavenumber < 1500 & ftir_data$wavenumber > 1000,
]

# First, fit the peaks (using the default 'voigt' method)
fitted_voigt <- fit_peaks(ftir_data, method = "voigt")

# Now, convert the fitted model object to a data frame
peak_df_voigt <- fit_peak_df(fitted_voigt)

print("Peak Data Frame from Voigt Fit:")
print(peak_df_voigt)
}
