% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak-fit.R
\name{get_fit_method}
\alias{get_fit_method}
\title{Get Fit Method}
\usage{
get_fit_method(fitted_peaks)
}
\arguments{
\item{fitted_peaks}{An object from [fit_peaks()].

  Un objet de [fit_peaks()].}
}
\value{
A character value for the peak type fitted to the spectra.

  Une valeur de caractère pour le type de pic ajusté aux spectres.
}
\description{
Determine the types of peaks used to create the [fit_peaks()]
  object.

  Déterminez les types de pics utilisés pour créer l'objet [fit_peaks()].
}
\keyword{internal}
