% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_plot_sample_ids}
\alias{get_plot_sample_ids}
\title{Get Plot Sample IDs}
\usage{
get_plot_sample_ids(ftir_spectra_plot)
}
\arguments{
\item{ftir_spectra_plot}{A plot generated by [plot_ftir()] or
  [plot_ftir_stacked()].

  Un tracé généré par [plot_ftir()] ou [plot_ftir_stacked()].}
}
\value{
A vector of factors corresponding to the sample IDs in the plot.

  unvecteur de facteurs correspondant aux ID d'échantillon dans le tracé
}
\description{
Get the sample IDs from a prepared plot. Useful if renaming in
  the plot legend.

  Obtenez les ID d’échantillon à partir d’un tracé préparé. Utile si vous
  renommez dans la légende de le tracé
}
\examples{
if (requireNamespace("ggplot2", quietly = TRUE)) {
  # Prepare a plot
  p <- plot_ftir(biodiesel)

  # Get the Sample IDs
  get_plot_sample_ids <- (p)
}
}
\seealso{
[rename_plot_sample_ids()]
}
