% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{intensity_type}
\alias{intensity_type}
\title{Intensity Type}
\usage{
intensity_type(ftir)
}
\arguments{
\item{ftir}{A data.frame in long format with columns `sample_id`, `wavenumber`,
and `absorbance`/`transmittance`.
Un data.frame au format long avec les colonnes `sample_id`, `wavenumber`,
et `absorbance`/`transmittance`.}
}
\value{
a character value 'absorbance' or 'transmittance'
}
\description{
Determines if the provided data has intensity type of absorbance
  or transmittance.
}
\keyword{internal}
