% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{is.ftir}
\alias{is.ftir}
\title{Is FTIR?}
\usage{
is.ftir(x)
}
\arguments{
\item{x}{object to test for type `ftir`}
}
\value{
True / False on if the item is `ftir`
}
\description{
Test if an object inherits class `ftir`
}
