% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{mean.ftir}
\alias{mean.ftir}
\title{Average together spectra}
\usage{
\method{mean}{ftir}(x, ...)
}
\arguments{
\item{x}{FTIR spectra data.frame of `class(x) == "ftir"`}

\item{...}{Additional parameters to supply to the underlying `average_spectra()` function.}
}
\value{
A data.frame containing the averaged FTIR spectra, with \code{sample_id}
corresponding to the provided \code{average_id}.

Un data.frame contenant les spectres IRTF moyennés, avec \code{sample_id}
correspondant à l'identifiant \code{average_id} fourni.
}
\description{
Calculates an average of two or more spectra.

Calcule la moyenne de deux spectres ou plus.
}
\seealso{
[average_spectra()]
}
