% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ftir.R
\name{plot_ftir}
\alias{plot_ftir}
\title{Plot FTIR}
\usage{
plot_ftir(ftir, plot_title = "FTIR Spectra", legend_title = "Sample ID")
}
\arguments{
\item{ftir}{A data.frame in long format with columns `sample_id`, `wavenumber`, `absorbance`
The `absorbance` column may be replaced by a `transmittance` column for transmittance plots.}

\item{plot_title}{A title for a plot. Defaults to "FTIR Spectra"}

\item{legend_title}{A title for the legend. Defaults to "Sample ID"}
}
\value{
a ggplot2 object
}
\description{
Plot the FTIR spectra in a journal prepared format. This is a baseplot only
}
