% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFTIR-package.R
\docType{package}
\name{PlotFTIR-package}
\alias{PlotFTIR}
\alias{PlotFTIR-package}
\title{PlotFTIR: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
