% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{+.ftir}
\alias{+.ftir}
\title{Addition of FTIR objects}
\usage{
\method{+}{ftir}(x, y)
}
\arguments{
\item{x}{First item to add. Either `x` or `y` must be an object of `class() == "ftir"`, and `y` or `x` must be scalar}

\item{y}{Second item to add. Either `x` or `y` must be an object of `class() == "ftir"`, and `y` or `x` must be scalar}
}
\value{
A data.frame containing the adjusted FTIR spectra.

  Un data.frame contenant les spectres IRTF ajustee.
}
\description{
Add a numeric (scalar) value to a FTIR spectra.
}
\seealso{
[add_scalar_value()]
}
