% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maths.R
\name{remove_continuum_ftir}
\alias{remove_continuum_ftir}
\title{Remove Continuum from FTIR Spectra}
\usage{
remove_continuum_ftir(ftir, type = "spline", application = "subtraction", ...)
}
\arguments{
\item{ftir}{A data.frame in long format with a single FTIR spectra in columns
  `sample_id`, `wavenumber`, and `absorbance` or `transmittance`.

  Un data.frame au format long avec un seul spectre IRTF dans les colonnes
  `sample_id`, `wavenumber`, et `absorbance` ou `transmittance`.}

\item{type}{The type of interpolation to use for the continuum. Options are
  `spline` (default) or `linear`.

  Le type d'interpolation à utiliser pour le continuum. Les options sont
  `spline` (par défaut) ou `linear`.}

\item{application}{How to apply the continuum to the spectra. Options are
  `subtraction` (default) or `division`.

  Comment appliquer le continuum aux spectres. Les options sont `subtraction`
  (par défaut) ou `division`.}

\item{...}{Additional arguments (currently unused).

  Arguments supplémentaires (actuellement inutilisés).}
}
\value{
A data.frame with the continuum removed from the spectra. The
  `absorbance` or `transmittance` column will be modified.

  Un data.frame avec le continuum supprimé des spectres. La colonne
  `absorbance` ou `transmittance` sera modifiée.
}
\description{
This function removes the continuum from FTIR spectra using
  either spline or linear interpolation. The continuum is defined as the
  convex hull of the spectrum, and is either subtracted or divided from the
  original spectrum. This is a common preprocessing step in reflectance
  spectroscopy to highlight absorption features.

  Cette fonction supprime le continuum des spectres IRTF en utilisant une
  interpolation spline ou linéaire. Le continuum est défini comme l'enveloppe
  convexe du spectre, et est soit soustrait, soit divisé du spectre original.
  Il s'agit d'une étape de prétraitement courante en spectroscopie de
  réflectance pour mettre en évidence les caractéristiques d'absorption.
}
\examples{
# Load the isopropanol sample spectrum
ftir_data <- sample_spectra[
  sample_spectra$sample_id == "isopropanol",
]

# Remove the continuum using spline interpolation and subtraction
ftir_no_continuum <- remove_continuum_ftir(ftir_data)

# Remove the continuum using linear interpolation and division
ftir_no_continuum_linear_div <- remove_continuum_ftir(
  ftir_data,
  type = "linear",
  application = "division"
)
}
\references{
Clark, R.N. and Roush, T.L. (1984) J. Geophysical Res. 89 pp
  6329-6340 <doi:10.1029/JB089iB07p06329>
}
