% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maths.R
\name{smooth_ftir}
\alias{smooth_ftir}
\title{Smooth FTIR with Savitzky-Golay filter}
\usage{
smooth_ftir(ftir, polynomial = 2, points = 13, derivative = 0)
}
\arguments{
\item{ftir}{A data.frame in long format with a single FTIR spectra in columns
  `sample_id`, `wavenumber`, and `absorbance`. The `absorbance` column may be
  replaced by a `transmittance` column for transmittance plots.

  Un data.frame au format long avec un seul spectre IRTF dans les colonnes
  `sample_id`, `wavenumber`, et `absorbance`. La colonne `absorbance` peut
  être remplacée par une colonne `transmittance` pour les tracés de
  transmittance.}

\item{polynomial}{Savitzky-Golay polynomial term.

  Terme polynomial de Savitzky-Golay.}

\item{points}{Savitzky-Golay points term.

  Terme de points de Savitzky-Golay}

\item{derivative}{Which derivative to return (default = 0 to smooth spectrum,
  but can alos determine smoothed derivatives)

  Dérivée à retourner (par défaut = 0 pour lisser le spectre, mais on peut
  aussi déterminer des dérivées lissées)}
}
\value{
a data.frame with an FTIR spectrum, smoothed (or the derivative)

  un data.frame avec un FTIR lissé, lissé (ou la dérivée lissée)
}
\description{
Smooth FTIR with Savitzky-Golay filter
}
\examples{
# Load the isopropanol sample spectrum
ftir_data <- sample_spectra[
  sample_spectra$sample_id == "isopropanol",
]

# Apply smoothing
ftir_smoothed <- smooth_ftir(ftir_data)
# --- Optional: Visualize the results ---
\dontrun{
  plot_ftir(ftir_smoothed, plot_title = "Smoothed FTIR")
}
}
\references{
* Savitzky, A.; Golay, M.J.E. (1964). "Smoothing and Differentiation of Data by Simplified Least Squares Procedures". Analytical Chemistry 36. pp. 1627–1639. doi:10.1021/ac60214a047
}
\seealso{
[signal::sgolayfilt()]
}
