% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{summary.ftir}
\alias{summary.ftir}
\title{Get a summary of FTIR class objects}
\usage{
\method{summary}{ftir}(x, ...)
}
\arguments{
\item{x}{the ftir class object to summarize}

\item{...}{additional parameters to print function.}
}
\value{
invisible, the passed in object
}
\description{
Get a summary of FTIR class objects
}
