% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulations.R
\name{compress_trans}
\alias{compress_trans}
\title{Compression Transformation}
\usage{
compress_trans(intercept = 2000, ratio = 5)
}
\arguments{
\item{intercept}{the wavenumber at which compression starts
le nombre d'onde auquel la compression commence}

\item{ratio}{the rate of compression (larger numbers are more compressed)
le taux de compression (des nombres plus grands sont plus compressés)}
}
\description{
used in `compress_low_energy()`
utilisé en `compress_low_energy`()`
}
\references{
From https://stackoverflow.com/a/43321639 & https://stackoverflow.com/a/78360580
}
\keyword{internal}
