% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_handling.R
\name{.get_language}
\alias{.get_language}
\title{Get current language setting}
\usage{
.get_language()
}
\value{
character string with the current language code
}
\description{
Get current language setting
}
\keyword{internal}
