% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{intensity_type}
\alias{intensity_type}
\title{Intensity Type}
\usage{
intensity_type(ftir)
}
\arguments{
\item{ftir}{A data.frame of FTIR spectral data including column to be
  converted. Can't contain both `absorbance` and `transmittance` column as
  the receiving column would be overwritten

  Un data.frame de données spectrales IRTF incluant la colonne à convertir.
  Ne peut pas contenir les colonnes `absorbance` et `transmittance` car la
  colonne de réception serait écrasée.}
}
\value{
a character value 'absorbance' or 'transmittance'
}
\description{
Determines if the provided data has intensity type of absorbance
  or transmittance.
}
\keyword{internal}
