% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ftir.R
\name{plot_ftir}
\alias{plot_ftir}
\title{Plot FTIR Spectra Overlaid}
\usage{
plot_ftir(ftir, plot_title = "FTIR Spectra", legend_title = "Sample ID")
}
\arguments{
\item{ftir}{A data.frame in long format with columns `sample_id`,
`wavenumber`, `absorbance` The `absorbance` column may be replaced by a
`transmittance` column for transmittance plots. The code determines the correct
y axis units and labels the plot/adjusts the margins appropriately.}

\item{plot_title}{A title for a plot. Defaults to "FTIR Spectra"}

\item{legend_title}{A title for the legend. Defaults to "Sample ID"}
}
\value{
a ggplot2 object
}
\description{
Produce a basic spectra overlay plot for all samples found in the ftir dataset provided.
}
\examples{
# Plot a basic FTIR Spectra overlay from the `sample_spectra` dataset with default titles
plot_ftir(sample_spectra)

}
