% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ftir.R
\name{plot_ftir_core}
\alias{plot_ftir_core}
\title{PlotFTIR core plot generator}
\usage{
plot_ftir_core(ftir, plot_title = "FTIR Spectra", legend_title = "Sample ID")
}
\arguments{
\item{ftir}{A data.frame in long format with columns `sample_id`,
 `wavenumber`, and `absorbance`. The `absorbance` column may be replaced by a
 `transmittance` column for transmittance plots. The code determines the
 correct y axis units and labels the plot/adjusts the margins appropriately.

 Un data.frame au format long avec les colonnes `sample_id`, `wavenumber`, et
 `absorbance`. La colonne `absorbance` peut être remplacée par une colonne
 `transmittance` pour les tracés de transmission. Le code détermine les
 unités correctes de l'axe Y et étiquette le tracé/ajuste les marges de
 manière appropriée.}

\item{plot_title}{A title for a plot. Defaults to "FTIR Spectra".

 Un titre pour une trace. La valeur par défaut est «FTIR Spectra».}

\item{legend_title}{A title for the legend. Defaults to "Sample ID".

 Un titre pour la légende. La valeur par défaut est «Sample ID».}
}
\value{
a ggplot object containing a  FTIR spectral plot. The plot and legend
 titles are as provided, with each sample provided a different default color
 from ggplot2.

 un objet ggplot contenant un tracé spectral FTIR. Les titres de le tracé
 et de la légende sont tels que fournis, chaque échantillon étant doté d'une
 couleur par défaut différente de celle de ggplot2.
}
\description{
Plot the FTIR spectra in a journal prepared format. This is the
 core plot code, please call [plot_ftir()] for basic (overlaid) plots and
 [plot_ftir_stacked()] for stacked and offset plots.

 Tracez les spectres FTIR dans un format préparé par un journal. C'est le
 code de tracé principal, veuillez appeler [plot_ftir()] pour les tracés de
 base (superposés) et [plot_ftir_stacked()] pour les tracés empilés et
 décalés.
}
