% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFTIR-package.R
\docType{package}
\name{PlotFTIR-package}
\alias{PlotFTIR}
\alias{PlotFTIR-package}
\title{PlotFTIR: Plot FTIR Spectra}
\description{
The goal of 'PlotFTIR' is to easily and quickly kick-start the production of journal-quality Fourier Transform Infra-Red (FTIR) spectral plots in R using 'ggplot2'. The produced plots can be published directly or further modified by 'ggplot2' functions. L'objectif de 'PlotFTIR' est de démarrer facilement et rapidement la production des tracés spectraux de spectroscopie infrarouge à transformée de Fourier (IRTF) de qualité journal dans R à l'aide de 'ggplot2'. Les tracés produits peuvent être publiés directement ou modifiés davantage par les fonctions 'ggplot2'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/NRCan/PlotFTIR}
  \item Report bugs at \url{https://github.com/NRCan/PlotFTIR/issues}
}

}
\author{
\strong{Maintainer}: Philip Bulsink \email{philip.bulsink@nrcan-rncan.gc.ca} (\href{https://orcid.org/0000-0001-9668-2429}{ORCID})

Other contributors:
\itemize{
  \item Ulrich Makanda \email{ulrich.makanda@nrcan-rncan.gc.ca} (Translated documentation to French/Traduire la documentation en français) [translator]
  \item His Majesty the King in Right of Canada, as represented by the Minister of Natural Resources [copyright holder]
}

}
\keyword{internal}
