% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_spectra}
\alias{sample_spectra}
\title{FTIR example data}
\format{
The data is in long format, ready for use, and as such has the following variables:

\describe{
  \item `wavenumber` wavenumber (x axis) value for each point. In units of wavenumbers (cm^-1)
  \item `absorbance` The absorbance of the sample at each specified wavenumber
  \item `sample_id` Sample ID of the various included spectra. See below for descriptions
}
}
\usage{
sample_spectra
}
\description{
This dataset hold example data from an ATR type FTIR instrument. 


The samples included in this data set are:
\itemize{
 \item `toluene` a FTIR spectra of toluene
 \item `ethanol` a FTIR spectra of ethanol
 \item `diesel` a FTIR spectra of a commercially available B0 diesel fuel from the Canadian market
}
}
\keyword{datasets}
